feat(Hosts): adds retry functions to check health/access to an OpenSearch cluster.
This adds a simple function that will request /_cluster/health with a linear backoff. The functions are recursive which means that the number of attempts will be distributed according to the formulat pow(2,n)-1, with default n=3, with a default delay off 5s.

---------

Signed off by: Simon Olander (simon.olander@sap.com)