feat(Hosts): implements failover mechanism to Hosts selection object
This modifies the original logic with attempting to choose multiple auth objects (username, password) to check which set of credentials yields an OK response. The logic will block the main thread, which is desired as we need to ensure that the creds are correct before continuing.

---------

Signed off by: Simon Olander (simon.olander@sap.com)